webapp/weather_station/requirements.txt
asgiref==3.7.2 ; python_version >= "3.11" and python_version < "4.0"
async-timeout==4.0.3 ; python_version >= "3.11" and python_full_version <= "3.11.2"
attrs==23.1.0 ; python_version >= "3.11" and python_version < "4.0"
autobahn==23.6.2 ; python_version >= "3.11" and python_version < "4.0"
automat==22.10.0 ; python_version >= "3.11" and python_version < "4.0"
cffi==1.15.1 ; python_version >= "3.11" and python_version < "4.0"
channels-redis==4.1.0 ; python_version >= "3.11" and python_version < "4.0"
channels==4.0.0 ; python_version >= "3.11" and python_version < "4.0"
constantly==15.1.0 ; python_version >= "3.11" and python_version < "4.0"
cryptography==41.0.4 ; python_version >= "3.11" and python_version < "4.0"
daphne==4.0.0 ; python_version >= "3.11" and python_version < "4.0"
django-bootstrap5==23.3 ; python_version >= "3.11" and python_version < "4.0"
django-filter==23.3 ; python_version >= "3.11" and python_version < "4.0"
django-leaflet==0.29.0 ; python_version >= "3.11" and python_version < "4.0"
django-rest-framework==0.1.0 ; python_version >= "3.11" and python_version < "4.0"
django==4.2.5 ; python_version >= "3.11" and python_version < "4.0"
djangorestframework==3.14.0 ; python_version >= "3.11" and python_version < "4.0"
hyperlink==21.0.0 ; python_version >= "3.11" and python_version < "4.0"
idna==3.4 ; python_version >= "3.11" and python_version < "4.0"
incremental==22.10.0 ; python_version >= "3.11" and python_version < "4.0"
leaflet==0.0.3 ; python_version >= "3.11" and python_version < "4.0"
mqttasgi==1.2.3 ; python_version >= "3.11" and python_version < "4.0"
msgpack==1.0.6 ; python_version >= "3.11" and python_version < "4.0"
paho-mqtt==1.6.1 ; python_version >= "3.11" and python_version < "4.0"
psycopg2==2.9.7 ; python_version >= "3.11" and python_version < "4.0"
pyasn1-modules==0.3.0 ; python_version >= "3.11" and python_version < "4.0"
pyasn1==0.5.0 ; python_version >= "3.11" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.11" and python_version < "4.0"
pyopenssl==23.2.0 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
pytz==2023.3.post1 ; python_version >= "3.11" and python_version < "4.0"
redis==5.0.0 ; python_version >= "3.11" and python_version < "4.0"
service-identity==23.1.0 ; python_version >= "3.11" and python_version < "4.0"
setuptools==68.2.2 ; python_version >= "3.11" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.11" and python_version < "4.0"
sqlparse==0.4.4 ; python_version >= "3.11" and python_version < "4.0"
twisted-iocpsupport==1.0.4 ; python_version >= "3.11" and python_version < "4.0" and platform_system == "Windows"
twisted[tls]==23.8.0 ; python_version >= "3.11" and python_version < "4.0"
txaio==23.1.1 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.8.0 ; python_version >= "3.11" and python_version < "4.0"
tzdata==2023.3 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "win32"
zope-interface==6.0 ; python_version >= "3.11" and python_version < "4.0"
